experiment with vector fields
In paraview, a vector field must have 3 dimensions to
display as glyph. If we have a 2D vector field, we have
to compute the 3D vector field with the calculator filter
using formula: e_X*iHat+e_Y*jHat

We can use SetLevelOfDetail to output interpolated values
from a high order elements. It seems that SetHighOrderOutput
doesn't have the same effect.

The output points corresponding to the level of details
selected do not match the quadrature location, and that is
why an interpolation is done when saving the grid for Paraview.

Signed-off-by: Giraldeau, Francis <giraldeau.francis@uqam.ca>